requirements.txt
langchain
openai
requests
python-dotenv
wikipedia
langchain-openai